test: relocate blaise api tests
